DB-3852: Performing batch reads from HBase.
HBase has a mechanism where you can set the maximum number of
cells to be returned by any one scan operation. This is necessary if we
are to avoid performing IO on cells which we don't care about, and also
if we are to have any control over our overall memory usage during
scans.

This introduces the "MergingReader", which uses the hbase batch api when
scanning data.