DB-3853: Introducing VACUUM_TABLE.
The Vacuum stored procedure can be invoked by calling call
SYSCS_UTIL.VACUUM_TABLE(<schema>,<table>); this will delete all records
< the minimum active transaction (at the moment), and will write a
checkpoint row in to the base table.

The purpose of the VACUUM_TABLE operation as a manual step (as opposed
to waiting for a compaction or automatic checkpointing) is to allow
users to disable compactions and still perform the necessary regular
maintenance. In most situations, this will be unused.