DB-3963: Adding commit timestamp to Checkpoints.
When performing read resolution, we want to add a commit timestamp to
the row. In most cases, this is its own cell, but for checkpoints this
isn't very useful because you'll see the checkpoint cell first. This
fixes that by adding the commit timestamp as the value in the checkpoint
cell instead of a commit timestamp.